Font rendering with outlines
Currently there is a problem when the line width is too big, the outline overflows
on the fill area. This requires the use of FT_Stroker but this will be in a separate commit.
For now, the feature is usable for small text with an outline of 2px.

The dots on i are rendered incorrectly, a fix is needed in lv_freetype_outline.c